fix layer->layer->component
