+++ built/core/builder/children.js
@@ -1,46 +1,11 @@
 "use strict";
-var __assign = (this && this.__assign) || function () {
-    __assign = Object.assign || function(t) {
-        for (var s, i = 1, n = arguments.length; i < n; i++) {
-            s = arguments[i];
-            for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p))
-                t[p] = s[p];
-        }
-        return t;
-    };
-    return __assign.apply(this, arguments);
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 var index_1 = require("../helper/index");
 var index_2 = require("../linter/index");
 var type_1 = require("../tsType/type");
-var error_1 = require("../error/error");
-var errorMessage_1 = require("../error/errorMessage");
 var type_2 = require("../tsType/type");
 var index_3 = require("../builder/index");
-var recursiveCheckFunctionAnswer = function (node) {
-    var haveDop = false;
-    var functionObject = {};
-    if (node["props"] !== undefined) {
-        functionObject = __assign({}, node["props"]);
-        haveDop = true;
-    }
-    if (node["child"] !== undefined) {
-        functionObject["children"] = node['child'];
-        haveDop = true;
-    }
-    var completeFunction = haveDop
-        ? node["tag"].bind(this)(__assign({}, functionObject))
-        : node["tag"].bind(this)();
-    var typeCompleteFunction = (0, index_1.typeOf)(completeFunction);
-    if (typeCompleteFunction !== "object") {
-        (0, error_1.default)("error ".concat(errorMessage_1.default.functionInTagReturn));
-    }
-    if (typeof completeFunction["tag"] === "function") {
-        return recursiveCheckFunctionAnswer.bind(this)(completeFunction);
-    }
-    return completeFunction;
-};
+var recuriveFunction_1 = require("./recuriveFunction");
 var recursiveChild = function (nodeProps, nodeChilds) {
     var _this = this;
     if (nodeProps === void 0) { nodeProps = null; }
@@ -68,7 +33,7 @@ var recursiveChild = function (nodeProps, nodeChilds) {
                 }
                 (0, index_2.validatorTagNode)(child);
                 if (typeof child["tag"] === "function") {
-                    var nodeTag = recursiveCheckFunctionAnswer.bind(_this)(child);
+                    var nodeTag = recuriveFunction_1.default.bind(_this)(child);
                     if (nodeTag["child"] !== undefined) {
                         nodeTag["child"] = recursiveChild.bind(_this)(nodeTag["props"], nodeTag["child"]);
                     }
@@ -100,7 +65,8 @@ var recursiveChild = function (nodeProps, nodeChilds) {
                     }
                     (0, index_2.validatorTagNode)(completeFunction);
                     if (typeof completeFunction["tag"] === "function") {
-                        completeFunction = recursiveCheckFunctionAnswer.bind(_this)(completeFunction);
+                        completeFunction = recuriveFunction_1.default.bind(_this)(completeFunction);
+                        console.log(completeFunction);
                     }
                     if (completeFunction["child"] !== undefined) {
                         completeFunction["child"] = recursiveChild.bind(_this)(completeFunction["props"], completeFunction["child"]);
@@ -131,7 +97,7 @@ var recursiveChild = function (nodeProps, nodeChilds) {
                 if (typeProxy === type_2.ProxyType.proxyComponent) {
                     var result = index_3.builder.bind(_this)(child.value);
                     if (typeof result["tag"] === "function") {
-                        result = recursiveCheckFunctionAnswer.bind(_this)(result);
+                        result = recuriveFunction_1.default.bind(_this)(result);
                     }
                     return {
                         type: type_1.Type.ProxyComponent,
