+++ src/core/builder/index.ts
@@ -9,43 +9,7 @@ import { Type } from "../tsType/type";
 import { Node, VNode } from "../tsType";
 
 import reqChild from "./children";
-
-interface Props {
-  children?: Array<Node>
-}
-
-const recursive = function(node: Node) {
-  let haveDop = false;
-  let propsCh : Props = {};
-
-  const { tag, props, child }: Node = node;
-
-  if (props !== undefined) {
-    propsCh = props;
-    haveDop = true;
-  }
-
-  if (child !== undefined) {
-    propsCh["children"] = child.flat(1);
-    haveDop = true;
-  }
-
-  const fTag : any = haveDop 
-    ? tag.bind(this)(propsCh) 
-    : tag.bind(this)();
-  
-  if (typeOf(fTag) !== "object") {
-    error(`rec-func - ${errorMessage.functionInTagReturn}`);
-  }
-
-  validatorTagNode(fTag);
-
-  if (typeof fTag["tag"] === "function") {
-    return recursive.bind(this)(fTag);
-  }
-
-  return fTag;
-}
+import recursive from "./recuriveFunction";
 
 export const builder = function(app: () => Node) : VNode {
   if (typeOf(app) !== "function") {
